canopy_fisheye {hemispheR}	R Documentation

Calculate canopy attributes from angular gap fraction data derived from fisheye images

Description

The function calculate canopy attributes from angular distribution of gap fraction. It returns both the effective (Le) and actual (L) leaf area index following the Miller theorem (1967). The Lang and Xiang (1986) clumping index LX is calculated as the ratio of Le to L; two additional clumping indices (LXG1, LXG2) are derived from ordered weighted average gap fraction as in Chianucci et al. 2019. The mean leaf angle (MTA) and the ellipsoidal x are derived from Norman and Campbell (1989). Canopy openness is also provided as weighted diffuse non-interceptance (DIFN), following the LAI-2200 manual.

Arguments

rfdw
Dataframe. The input dataframe generated from gapfrac_fisheye(), which contains gap fraction for zenith and azimuth bins.

Value

A dataframe of canopy attributes from classified fisheye images.

See Also

Chianucci F., Zou J., Leng P., Zhuang Y., Ferrara C. 2019. A new method to estimate clumping index integrating gap fraction averaging with the analysis of gap size distribution. Canadian Journal of Forest Research 49 https://doi.org/10.1139/cjfr-2018-0213.

LAI-2200C Plant Canopy Analyzer - Instruction Manuals. Licor https://www.licor.com/env/support/LAI-2200C/manuals.html.

Lang A.R.G., Xiang Y. 1986. Estimation of leaf area index from transmission of direct sunlight in discontinuous canopies. Agricultural and Forest Meteorology 37, 228-243. https://doi.org/10.1016/0168-1923(86)90033-X.

Miller J.B. 1967. A formula for average foliage density. Australian Journal of Botany 15(1) 141 - 144. https://www.publish.csiro.au/BT/BT9670141.

Norman J.M., Campbell G.S. 1986. Canopy structure. In: Plant Physiological Ecology, pp. 301-325 https://link.springer.com/chapter/10.1007/978-94-009-2221-1_14.

Examples

c.im<-system.file('extdata/circular_coolpix4500+FC-E8_chestnut.jpg',package='hemispheR')
c.im %>%
  import_fisheye(circ.mask=camera_fisheye('Coolpix4500+FC-E8')) %>%
  binarize_fisheye() %>%
  gapfrac_fisheye(lens='FC-E8',nrings=7,nseg=8,endVZA=70,display=TRUE) %>%
  canopy_fisheye()

#Zenith rings similar to LAI-2000/2200:
c.im %>%
 import_fisheye(circ.mask=camera_fisheye('Coolpix4500+FC-E8')) %>%
  binarize_fisheye() %>%
  gapfrac_fisheye(lens='FC-E8',nrings=5,nseg=8,endVZA=75,display=TRUE) %>%
  canopy_fisheye()

#The hinge angle method close to 1 radian (57 degree):
c.im %>%
 import_fisheye(circ.mask=camera_fisheye('Coolpix4500+FC-E8')) %>%
  binarize_fisheye() %>%
  gapfrac_fisheye(lens='FC-E8',nrings=1,nseg=8,startVZA=55,endVZA=60,display=TRUE) %>%
  canopy_fisheye()

[Package hemispheR version 0.0.0.9000 Index]