zonal_mask {hemispheR}	R Documentation

Divide a raster image into four stacks which are used as masks.

Description

This function imports a raster image using raster::raster() functionality, and divide into four masks, using the image centre and borders as vertices. The four zonal masks are then returned as a raster stack.

Usage

zonal_mask(img)

Arguments

img
Raster. The input raster image

Value

A 4-layers stacks of image masks

Examples

image<-system.file('extdata/circular_coolpix4500+FC-E8_chestnut.jpg',package='hemispheR')
zmsk<-zonal_mask(raster::raster(image))
raster::plot(zmsk,col=gray.colors(5),main=c('N','W','S','E'))

[Package hemispheR version 0.0.0.9000 Index]